chore: [GSW-2004] Moved the Pages folder under the `/src` folder
