+++ packages/web/src/pages/404.tsx
