+++ packages/web/src/pages/500.tsx
