+++ packages/web/src/pages/earn/pool/remove.tsx
