+++ packages/web/src/pages/earn/pool/stake.tsx
