+++ packages/web/src/pages/privacy/index.tsx
