+++ packages/web/src/pages/terms/index.tsx
